Relocated exceptions, tweaked phpunit config, fixed test
