+++ phpunit.xml.dist
@@ -3,6 +3,10 @@
 <phpunit backupGlobals="false"
          backupStaticAttributes="false"
          colors="true"
+         verbose="true"
+         beStrictAboutTestsThatDoNotTestAnything="true"
+         beStrictAboutOutputDuringTests="true"
+         beStrictAboutTestSize="true"
          convertErrorsToExceptions="true"
          convertNoticesToExceptions="true"
          convertWarningsToExceptions="true"
